refactor: delete event store service shell
